[Enhancement][Olana] Moved components into their own files
